remove next from src, only use it for lint
