+++ jest.config.js
@@ -1,14 +1,7 @@
-const nextJest = require("next/jest");
-
-const createJestConfig = nextJest({
-    // next.config.jsとテスト環境用の.envファイルが配置されたディレクトリをセット。基本は"./"で良い。
-    dir: "./",
-});
-
 // Jestのカスタム設定を設置する場所。従来のプロパティはここで定義。
 const customJestConfig = {
     // jest.setup.jsを作成する場合のみ定義。
-    // setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
+    setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
     moduleNameMapper: {
         // aliasを定義 （tsconfig.jsonのcompilerOptions>pathsの定義に合わせる）
         "^@/components/(.*)$": "<rootDir>/components/$1",
@@ -18,5 +11,4 @@ const customJestConfig = {
     roots: ["<rootDir>/tests/"],
 };
 
-// createJestConfigを定義することによって、本ファイルで定義された設定がNext.jsの設定に反映されます
-module.exports = createJestConfig(customJestConfig);
+module.exports = customJestConfig;
