+++ jest.setup.js
@@ -1,0 +1,12 @@
+import 'regenerator-runtime/runtime';
+
+jest.mock('react-polyglot', () => {
+  const originalModule = jest.requireActual('react-polyglot');
+  return {
+    __esModule: true,
+    ...originalModule,
+    useTranslate: jest.fn(() => {
+      return (key) => key;
+    }),
+  };
+});
