+++ package.json
@@ -32,6 +32,7 @@
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "eslint-plugin-unused-imports": "^2.0.0",
     "prettier": "^2.7.1",
+    "regenerator-runtime": "^0.13.11",
     "sass": "^1.54.5",
     "stylelint": "^14.11.0",
     "stylelint-config-css-modules": "^4.1.0",
