+++ tests/components/HistoricalVaultLogTable.test.jsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { prepareMocksResult } from '../utils/helper';
 import HistoricalVaultLogTable from '../../src/components/HistoricalVaultLogTable';
 
 describe('HistoricalVaultLogTable component', () => {
   beforeAll(() => {});
   test('should show table header', () => {
-    expect(prepareMocksResult).toBeTruthy();
     render(
       <table>
         <HistoricalVaultLogTable heading={true} />
