+++ tests/utils/helper.js
@@ -1,14 +1,2 @@
-const prepareMocks = () => {
-  jest.mock('react-polyglot', () => {
-    const originalModule = jest.requireActual('react-polyglot');
-    return {
-      __esModule: true,
-      ...originalModule,
-      useTranslate: jest.fn(() => {
-        return (key) => key;
-      }),
-    };
-  });
-  return true;
-};
-export const prepareMocksResult = prepareMocks();
+const helper = {};
+export default helper;
